v1.8.1 update
Fixed issue with shift+item
Added long messages, which allow you to send multiple messages with 1
hotkey.
Added Garena Autojoiner
Added option to autodetect (not compatible with window mode)
Replace sendplay to send mode for more reliable function the share
control and the whole script on old machines
removed wait variable (for share control)
Improved garena autojoiner (from now you don't need to deactivate it
when you join the room and increased time period to 5.010s)
added option ShopIndependence. If it's on - the inventory will not work
when you click on the shops => you can have same hotkeys on inventory
and items in the shop (so hard to explain....)